requirements.txt
MouseInfo==0.1.3
numpy==2.0.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
PyYAML==6.0.2
